python/requirements.txt
bottle==0.12.25
pyserial==3.5
